requirements.txt
google-genai
Pillow
rich
click
